pyproject.toml
[tool.poetry]
name = "test-task2"
version = "0.1.0"
description = ""
authors = ["Leonid3# <antonfenrisis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
